+++ lib/ngtcp2_rtb.h
@@ -191,9 +191,8 @@ void ngtcp2_rtb_lost_pop(ngtcp2_rtb *rtb);
  * NGTCP2_ERR_CALLBACK_FAILURE
  *     User callback failed
  */
-int ngtcp2_rtb_recv_ack(ngtcp2_rtb *rtb, const ngtcp2_ack *fr,
-                        uint8_t unprotected, ngtcp2_conn *conn,
-                        ngtcp2_tstamp ts);
+int ngtcp2_rtb_recv_ack(ngtcp2_rtb *rtb, const ngtcp2_ack *fr, int unprotected,
+                        ngtcp2_conn *conn, ngtcp2_tstamp ts);
 
 void ngtcp2_rtb_detect_lost_pkt(ngtcp2_rtb *rtb, ngtcp2_rcvry_stat *rcs,
                                 uint64_t largest_ack, uint64_t last_tx_pkt_num,
